Moved USWDSComboBox.vue to form components folder
